cpts<- {changepoint}	R Documentation

Generic Function - cpts<-

Description

Generic function

Usage

cpts(object)<-value

Arguments

object
Depending on the class of object depends on the method used (and if one exists)

value
Replacement value

Details

Generic function.

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

cpts<–methods

Examples

x=new("cpt") # new cpt object
cpts(x)<-10 # replaces the vector of changepoint in object x with 10

[Package changepoint version 2.2.1 Index]