cpts.ts {changepoint}	R Documentation

Generic Function - cpts.ts

Description

Generic function

Usage

cpts.ts(object)

Arguments

object
Depending on the class of object depends on the method used (and if one exists)

Details

Generic function.

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

cpts.ts-methods

Examples

x=new("cpt") # new cpt object
cpts.ts(x) # retrieves the cpts slot from x but unlike cpts, displays the "date" index rather than
#the position within the vector

[Package changepoint version 2.2.1 Index]