cpts.ts-methods {changepoint}	R Documentation

~~ Methods for Function cpts.ts ~~

Description

~~ Methods for function cpts.ts ~~

Methods

signature(object = "cpt")

Retrieves cpts slot from an object of class cpt, from version 1.0 this no longer prints the length of the dataset. Contrary to the cpts function, cpts.ts displays the changepoints using the "date" index according to the start and frequency of the time series data.set(object)

signature(object = "cpt.reg")

Retrieves cpts slot from an object of class cpt.reg, from version 1.0 this no longer prints the length of the dataset. Contrary to the cpts function, cpts.ts displays the changepoints using the "date" index according to the start and frequency of the time series data.set(object)

[Package changepoint version 2.2.1 Index]