distribution {changepoint}	R Documentation

Generic Function - distribution

Description

Generic function

Usage

distribution(object)

Arguments

object
Depending on the class of object depends on the method used (and if one exists)

Details

Generic Function

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

distribution-methods

Examples

x=new("cpt") # new cpt object
distribution(x) # retrieves the distribution slot from x

[Package changepoint version 2.2.1 Index]