likelihood {changepoint}	R Documentation

Generic Function - likelihood

Description

Generic function to calculate the likelihood

Usage

likelihood(object)

Arguments

object
Depending on the class of object depends on the method used to calculate the likelihood (and if one exists)

Details

Generic Function to calculate the likelihood.

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

likelihood-methods, cpt.mean,cpt.var,cpt.meanvar

Examples

set.seed(1)
x=c(rnorm(50,0,1),rnorm(50,0,10),rnorm(50,0,5),rnorm(50,0,1))
out=cpt.var(x,penalty="Manual",pen.value="2*log(n)",method="BinSeg",Q=5)
likelihood(out) # returns the raw scaled negative likelihood (925.8085) and the scaled negative
#likelihood + penalty (957.5984)

[Package changepoint version 2.2.1 Index]