param {changepoint}	R Documentation

Generic Function - param

Description

Generic function that returns parameter estimates.

Usage

param(object,...)

Arguments

object
Depending on the class of object depends on the method used to find the parameter estimates (and if one exists)

...
Other variables that may be required depending on the class of object, see individual methods.

Details

Generic Function that returns parameter estimates

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

param-methods,cpt.mean,cpt.var,cpt.meanvar

Examples

set.seed(1)
x=c(rnorm(100,0,1),rnorm(100,0,10))
ans=cpt.var(x,penalty="Asymptotic",pen.value=0.01,method="AMOC",param.estimates=FALSE)
# if estimates were not requested in the analysis then they can be created at a later stage if
#required
ans=param(ans) # fills the param.est slot with the parameter estimes.
param.est(ans) # variance is 0.8067621

[Package changepoint version 2.2.1 Index]