param-methods {changepoint}	R Documentation

~~ Methods for Function param ~~

Description

~~ Methods for function param ~~

Methods

signature(object = "cpt")

Replaces the param.value slot in object with the parameter estimates that are appropriate for the changepoint type (cpttype slot). If the Gamma test statistic is used then the shape parameter is required as a variable.

signature(object = "cpt.range")

Returns a new (blank) cpt.range object where the param.value slot contains the parameter estimates for a segmentation with the ncpts specified number of changepoints. This is also tailored to the changepoint type (cpttype slot) of the original object. If the Gamma test statistic is used then the shape parameter is required as a variable. An example of use is given in the cpt.mean help file.

signature(object = "cpt.reg")

Replaces the param.value slot in object with the parameter estimates that are appropriate for the changepoint test statistic (test.stat slot).

[Package changepoint version 2.2.1 Index]