param.est {changepoint}	R Documentation

Generic Function - param.est

Description

Generic function

Usage

param.est(object)

Arguments

object
Depending on the class of object depends on the method used (and if one exists)

Details

Generic Function

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

param.est-methods

Examples

x=new("cpt") # new cpt object
param.est(x) # retrieves the param.est slot from x

[Package changepoint version 2.2.1 Index]