param.est<- {changepoint}	R Documentation

Generic Function - param.est<-

Description

Generic function

Usage

param.est(object)<-value

Arguments

object
Depending on the class of object depends on the method used (and if one exists)

value
Replacement value

Details

Generic Function

Value

Depends on the class of object, see individual methods

Author(s)

Rebecca Killick

See Also

param.est<–methods

Examples

x=new("cpt") # new cpt object
param.est(x)<-list(mean=0) # replaces the current param.est list in x with list(mean=0)

[Package changepoint version 2.2.1 Index]