plot-methods {changepoint}	R Documentation

~~ Methods for Function plot in Package ‘graphics’ ~~

Description

~~ Methods for function plot in Package ‘graphics’ ~~

Methods

signature(x = "ANY")

Generic plot function, see graphics package description using ?plot

signature(x = "cpt")

Plots the data and identifies the changepoints using vertical lines (change in variance), horizontal lines (change in mean). Optional arguments to control the lines: cpt.col equivilent to col to change the colour of the changepoint line; cpt.width equivilent to lwd to change the width of the changepoint line; cpt.style equivilent to lty to change the style of the line.

signature(x = "cpt.range")

As for the cpt objects except for two optional arguments, ncpts and diagnostic. The ncpts option allows you to specify a plot of the segmentation with ncpts changepoints in, i.e. the optimal may be specified as 10 changes but you want to plot the segmentation with 5 changes (provided a segmentation with 5 changes is listed in cpts.full(x). The diagnostic option when set to TRUE plots the number of changepoints in each segmentation against the change in test statistic when adding that change. This can aide the decision on the number of changepoints as when a true changepoint is added the cost increases/decreases rapidly, but when a changepoint due to noise is added the change is small. This is akin to a scree plot in principal component analysis. The idea is that someone may choose to create a plot using diagnostic=TRUE, identify the appropriate number of changes and then replot using ncpts to visualize that segmentation.

signature(x = "cpt.reg")

Plotting is only valid for one regressor. Plots the regressor against the response and identifies the changepoints using horizontal lines. Optional arguments to control the lines: cpt.col equivilent to col to change the colour of the changepoint line; cpt.width equivilent to lwd to change the width of the changepoint line; cpt.style equivilent to lty to change the style of the line.

[Package changepoint version 2.2.1 Index]