seg.len-methods {changepoint}	R Documentation

~~ Methods for Function seg.len ~~

Description

~~ Methods for function seg.len ~~

Methods

signature(object = "cpt")

Returns the length of segments from an object of class cpt, i.e. the number of observations between changepoints

signature(object = "cpt.reg")

Returns the length of segments from an object of class cpt.reg, i.e. the number of observations between changepoints

[Package changepoint version 2.2.1 Index]